Move memory bank to models/modules
* This fixes cyclic imports between lightly(models/modules and lightly/losses
