+++ examples/pytorch_lightning/swav_queue.py
@@ -8,7 +8,7 @@ import torchvision
 from torch import nn
 
 from lightly.loss import SwaVLoss
-from lightly.loss.memory_bank import MemoryBankModule
+from lightly.models.modules.memory_bank import MemoryBankModule
 from lightly.models.modules import SwaVProjectionHead, SwaVPrototypes
 from lightly.transforms.swav_transform import SwaVTransform
