+++ lightly/models/modules/memory_bank.py
