+++ lightly/models/modules/nn_memory_bank.py
@@ -5,7 +5,7 @@
 
 import torch
 
-from lightly.loss.memory_bank import MemoryBankModule
+from lightly.models.modules.memory_bank import MemoryBankModule
 
 
 class NNMemoryBankModule(MemoryBankModule):
